requirements.txt
playwright==1.40.0
pydantic==2.5.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
lxml==4.9.3
beautifulsoup4==4.12.2
selenium==4.15.2
httpx==0.25.2
aiohttp==3.9.1
loguru==0.7.2
slack-sdk==3.26.1
python-json-logger==2.0.7
